OLD/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "racing-coach"
version = "0.1.0"
description = "Racing coach telemetry analysis for iRacing"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Sawyer", email = "sawyerczupka@gmail.com" }]
dependencies = [
    "pyirsdk>=1.3.5",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "pandas>=2.2.3",
    "fastparquet>=2024.11.0",
    "sqlalchemy>=2.0.39",
    "psycopg2-binary>=2.9.10",
]

[project.scripts]
run-coach = "racing_coach.core.app:main"

[project.optional-dependencies]
dev = ["black>=24.10.0", "isort>=5.13.2"]
notebook = ["plotly>=5.24.1", "ipykernel>=6.29.5", "nbformat>=5.10.4"]

[project.urls]
Repository = "https://github.com/yourusername/racing-coach"
